Added ability to activate/deactivate @everyone
You can now activate/deactivate the @everyone mention in the "content" field of the webhook, making it less intrusive. You can parameter this for each server, or you can change the default value which is true.

I modified a bit the JSON file . Any "comment" labeled field are not used in the webhook, and are just here to allow for better understanding.
I repeat that the JSON file included is not working, it is just supposed to help you understand how to make your own JSON file.